+++ p2n/tables.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 # (c) 2017 The Patent2Net Developers
 import logging
-from p2n.util import filterdict, dictproduct
+import operator
+from p2n.util import filterdict, dictproduct, unique
 
 logger = logging.getLogger(__name__)
 
@@ -19,7 +20,7 @@ def pivottables_data_brevets(document_list):
     scalar values (i.e. strings or numbers) and not objects, see also:
     https://github.com/nicolaskruchten/pivottable/wiki/Input-Formats
 
-    This is a clean-room reimplementation of the core of
+    This is an attempt of a clean-room reimplementation of the core of
     "FormateExportPivotTable.py" and "P2N_Lib.DecoupeOnTheFly".
     """
 
@@ -40,6 +41,91 @@ def pivottables_data_brevets(document_list):
 
         # Use only designated fields, skip others
         document = filterdict(document, fields)
+
+        # Compute cartesian product to satisfy PivotTable.js
+        expanded = dictproduct(document)
+
+        # Add to list of expanded data
+        expanded_all += expanded
+
+    return expanded_all
+
+
+def pivottables_data_documents(document_list):
+    """
+    Compute data suitable for feeding to PivotTable.js from OPSExchangeDocument data model.
+
+    Obtains a list of document objects of type OPSExchangeDocument, simplifies its structure and
+    filters each dictionary by designated keys and normalizes/flattens the dictionaries
+    by expanding all entries containing multivalued attributes to all combinations
+    of themselves with scalar values only (cartesian product).
+
+    The reason for this is because PivotTable.js assumes all attributes to be
+    scalar values (i.e. strings or numbers) and not objects, see also:
+    https://github.com/nicolaskruchten/pivottable/wiki/Input-Formats
+
+    This is an attempt of a clean-room reimplementation of the core of
+    "FormateExportPivotTable.py" and "P2N_Lib.DecoupeOnTheFly"
+    together with dynamic rule-based dictionary transformation
+    to massage the data appropriately.
+    """
+
+    rules = [
+
+        # Simple scalar fields not requiring any transformation
+        'document_number',
+        'country',
+        'kind',
+        'application_year',
+        'publication_year',
+
+        #'priority-active-indicator',   # TODO
+        #'prior-Date',                  # TODO
+        {
+            'name': 'ipcr4',
+            'getter': lambda document: document.classifications['IPCR'],
+            'recipe': lambda items: unique([item[:4] for item in items]),
+        },
+        {
+            'name': 'ipcr7',
+            'getter': lambda document: document.classifications['IPCR'],
+            'recipe': lambda items: unique([item.split('/')[0] for item in items]),
+        },
+        {
+            'name': 'applicant',
+            'getter': operator.attrgetter('applicants'),
+            'recipe': lambda items: unique([item['name'] for item in items if item]),
+        },
+        {
+            'name': 'inventor',
+            'getter': operator.attrgetter('inventors'),
+            'recipe': lambda items: unique([item['name'] for item in items if item]),
+        },
+        {
+            'name': 'applicant_country',
+            'getter': operator.attrgetter('applicants'),
+            'recipe': lambda items: unique([item['country'] for item in items if item]),
+        },
+        {
+            'name': 'inventor_country',
+            'getter': operator.attrgetter('inventors'),
+            'recipe': lambda items: unique([item['country'] for item in items if item]),
+        },
+        {
+            'name': 'designated_state',
+            'getter': lambda document: document.register and document.register.designated_states,
+        },
+
+        # Excluded due to excessive amount of entries after multivalue expansion
+        #"CPC",
+        #"equivalents",
+    ]
+
+    expanded_all = []
+    for document in document_list:
+
+        # Use only designated fields, skip others
+        document = document.transform_dict(rules)
 
         # Compute cartesian product to satisfy PivotTable.js
         expanded = dictproduct(document)